Moved CST next to ConcreteParser.
