+++ waspls/src/Wasp/LSP/Completion.hs
@@ -8,9 +8,9 @@ import Data.Maybe (maybeToList)
 import qualified Data.Text as Text
 import qualified Language.LSP.Types as LSP
 import qualified Language.LSP.Types.Lens as LSP
-import Wasp.Analyzer.Parser.ConcreteParser.CST (SyntaxNode)
-import qualified Wasp.Analyzer.Parser.ConcreteParser.CST as S
-import Wasp.Analyzer.Parser.ConcreteParser.CST.Traverse
+import Wasp.Analyzer.Parser.CST (SyntaxNode)
+import qualified Wasp.Analyzer.Parser.CST as S
+import Wasp.Analyzer.Parser.CST.Traverse
 import Wasp.LSP.ServerM
 import Wasp.LSP.ServerState
 import Wasp.LSP.Syntax (findChild, isAtExprPlace, lexemeAt, lspPositionToOffset, showNeighborhood, toOffset)
