+++ waspls/src/Wasp/LSP/Syntax.hs
@@ -14,8 +14,8 @@ where
 
 import Data.List (find, intercalate)
 import qualified Language.LSP.Types as J
-import qualified Wasp.Analyzer.Parser.ConcreteParser.CST as S
-import Wasp.Analyzer.Parser.ConcreteParser.CST.Traverse
+import qualified Wasp.Analyzer.Parser.CST as S
+import Wasp.Analyzer.Parser.CST.Traverse
 
 -- | @lspPositionToOffset srcString position@ returns 0-based offset from the
 -- start of @srcString@ to the specified line and column.
